+++ libraries/ScrewTheoryLib/ScrewTheoryIkProblem.cpp
@@ -3,129 +3,8 @@
 #include "ScrewTheoryIkProblem.hpp"
 
 #include <algorithm>
-#include <functional>
-#include <iterator>
-#include <set>
-
-#include "ScrewTheoryIkSubproblems.hpp"
 
 using namespace roboticslab;
-
-// -----------------------------------------------------------------------------
-
-namespace
-{
-    inline bool liesOnAxis(const MatrixExponential & exp, const KDL::Vector & p)
-    {
-        return KDL::Equal((p - exp.getOrigin()) * exp.getAxis(), KDL::Vector::Zero());
-    }
-
-    inline bool parallelAxes(const MatrixExponential & exp1, const MatrixExponential & exp2)
-    {
-        return KDL::Equal(exp1.getAxis() * exp2.getAxis(), KDL::Vector::Zero());
-    }
-
-    inline bool perpendicularAxes(const MatrixExponential & exp1, const MatrixExponential & exp2)
-    {
-        return KDL::Equal(KDL::dot(exp1.getAxis(), exp2.getAxis()), 0);
-    }
-
-    inline bool colinearAxes(const MatrixExponential & exp1, const MatrixExponential & exp2)
-    {
-        return parallelAxes(exp1, exp2) && liesOnAxis(exp1, exp2.getOrigin());
-    }
-
-    bool intersectingAxes(const MatrixExponential & exp1, const MatrixExponential & exp2, KDL::Vector & p)
-    {
-        // "Intersection of Two Lines in Three-Space" by Ronald Goldman, University of Waterloo (Waterloo, Ontario, Canada)
-        // published in: "Graphic Gems", edited by Andrew S. Glassner, 1 ed., ch. 5, "3D Geometry" (p. 304)
-        // referenced in: https://stackoverflow.com/a/565282
-
-        KDL::Vector cross = exp1.getAxis() * exp2.getAxis();
-        KDL::Vector diff = exp2.getOrigin() - exp1.getOrigin();
-
-        double den = KDL::pow(cross.Norm(), 2);
-        double t = KDL::dot(cross, diff * exp2.getAxis()) / den;
-        double s = KDL::dot(cross, diff * exp1.getAxis()) / den;
-
-        KDL::Vector L1 = exp1.getOrigin() + exp1.getAxis() * t;
-        KDL::Vector L2 = exp2.getOrigin() + exp2.getAxis() * s;
-
-        p = L1;
-
-        return KDL::Equal(L1, L2);
-    }
-
-    inline bool planarMovement(const MatrixExponential & exp1, const MatrixExponential & exp2)
-    {
-        bool sameMotionType = exp1.getMotionType() == exp2.getMotionType();
-        return sameMotionType ? parallelAxes(exp1, exp2) : perpendicularAxes(exp1, exp2);
-    }
-
-    inline bool normalPlaneMovement(const MatrixExponential & exp1, const MatrixExponential & exp2)
-    {
-        bool sameMotionType = exp1.getMotionType() == exp2.getMotionType();
-        return sameMotionType ? perpendicularAxes(exp1, exp2) : parallelAxes(exp1, exp2);
-    }
-
-    struct compare_vectors : public std::binary_function<const KDL::Vector &, const KDL::Vector &, bool>
-    {
-        result_type operator()(first_argument_type lhs, second_argument_type rhs)
-        {
-            if (KDL::Equal(lhs.x(), rhs.x()))
-            {
-                if (KDL::Equal(lhs.y(), rhs.y()))
-                {
-                    if (KDL::Equal(lhs.z(), rhs.z()))
-                    {
-                        // treat as equal per !comp(a, b) && !comp(b, a) == true
-                        // https://en.cppreference.com/w/cpp/container/set
-                        return false;
-                    }
-                    else if (lhs.z() < rhs.z())
-                    {
-                        return true;
-                    }
-                }
-                else if (lhs.y() < rhs.y())
-                {
-                    return true;
-                }
-            }
-            else if (lhs.x() < rhs.x())
-            {
-                return true;
-            }
-
-            return false;
-        }
-    };
-
-    class poe_term_candidate : public std::unary_function<const ScrewTheoryIkProblemBuilder::PoeTerm &, bool>
-    {
-    public:
-        poe_term_candidate(result_type _expectedKnown)
-            : expectedKnown(_expectedKnown),
-              expectedSimplified(false),
-              ignoreSimplifiedState(true)
-        {}
-
-        poe_term_candidate(result_type _expectedKnown, result_type _expectedSimplified)
-            : expectedKnown(_expectedKnown),
-              expectedSimplified(_expectedSimplified),
-              ignoreSimplifiedState(false)
-        {}
-
-        result_type operator()(argument_type poeTerm)
-        {
-            return poeTerm.known == expectedKnown && (ignoreSimplifiedState || poeTerm.simplified == expectedSimplified);
-        }
-
-    private:
-        result_type expectedKnown, expectedSimplified, ignoreSimplifiedState;
-    }
-    knownTerm(true), unknownTerm(false), unknownNotSimplifiedTerm(false, false);
-}
 
 // -----------------------------------------------------------------------------
 
@@ -144,88 +23,6 @@ ScrewTheoryIkProblem::~ScrewTheoryIkProblem()
         delete steps[i];
         steps[i] = NULL;
     }
-}
-
-// -----------------------------------------------------------------------------
-
-ScrewTheoryIkProblemBuilder::ScrewTheoryIkProblemBuilder(const PoeExpression & _poe)
-    : poe(_poe),
-      poeTerms(poe.size())
-{}
-
-// -----------------------------------------------------------------------------
-
-std::vector<KDL::Vector> ScrewTheoryIkProblemBuilder::searchPoints(const PoeExpression & poe)
-{
-    std::set<KDL::Vector, compare_vectors> set;
-
-    set.insert(KDL::Vector::Zero());
-
-    for (int i = 0; i < poe.size(); i++)
-    {
-        const MatrixExponential & exp = poe.exponentialAtJoint(i);
-
-        if (exp.getMotionType() != MatrixExponential::ROTATION)
-        {
-            continue;
-        }
-
-        set.insert(exp.getOrigin());
-
-        for (int j = 1; j < poe.size(); j++)
-        {
-            const MatrixExponential & exp2 = poe.exponentialAtJoint(j);
-
-            if (exp2.getMotionType() != MatrixExponential::ROTATION)
-            {
-                continue;
-            }
-
-            if (parallelAxes(exp, exp2))
-            {
-                continue;
-            }
-
-            if (colinearAxes(exp, exp2))
-            {
-                continue;
-            }
-
-            KDL::Vector p;
-
-            if (intersectingAxes(exp, exp2, p))
-            {
-                set.insert(p);
-            }
-        }
-    }
-
-    for (int i = 0; i < poe.size(); i++)
-    {
-        const MatrixExponential & exp = poe.exponentialAtJoint(i);
-
-        if (exp.getMotionType() != MatrixExponential::ROTATION)
-        {
-            continue;
-        }
-
-        double factor;
-        KDL::Vector randomPointOnAxis;
-
-        do
-        {
-            KDL::random(factor);
-            randomPointOnAxis = exp.getOrigin() + factor * exp.getAxis();
-        }
-        while (!set.insert(randomPointOnAxis).second);
-    }
-
-    set.insert(poe.getTransform().p);
-
-    std::vector<KDL::Vector> points(set.size());
-    std::copy(set.begin(), set.end(), points.begin());
-
-    return points;
 }
 
 // -----------------------------------------------------------------------------
@@ -447,341 +244,6 @@ ScrewTheoryIkProblem * ScrewTheoryIkProblem::create(const PoeExpression & poe, c
 {
     // TODO: validate
     return new ScrewTheoryIkProblem(poe, steps, reversed);
-}
-
-// -----------------------------------------------------------------------------
-
-ScrewTheoryIkProblem * ScrewTheoryIkProblemBuilder::build()
-{
-    // TODO: deallocate memory, implement destructor
-
-    std::vector<ScrewTheoryIkSubproblem *> steps = searchSolutions();
-
-    if (std::count_if(poeTerms.begin(), poeTerms.end(), knownTerm) == poe.size())
-    {
-        return ScrewTheoryIkProblem::create(poe, steps);
-    }
-
-    poe = poe.reverse();
-
-    for (std::vector<PoeTerm>::iterator it = poeTerms.begin(); it != poeTerms.end(); ++it)
-    {
-        it->known = false;
-    }
-
-    steps = searchSolutions();
-
-    if (std::count_if(poeTerms.begin(), poeTerms.end(), knownTerm) == poe.size())
-    {
-        return ScrewTheoryIkProblem::create(poe, steps, true);
-    }
-
-    for (std::vector<ScrewTheoryIkSubproblem *>::iterator it = steps.begin(); it != steps.end(); ++it)
-    {
-        delete (*it);
-        *it = NULL;
-    }
-
-    return NULL;
-}
-
-// -----------------------------------------------------------------------------
-
-std::vector<ScrewTheoryIkSubproblem *> ScrewTheoryIkProblemBuilder::searchSolutions()
-{
-    points = searchPoints(poe);
-
-    testPoints.assign(MAX_SIMPLIFICATION_DEPTH, points[0]);
-
-    std::vector<ScrewTheoryIkSubproblem *> steps;
-
-    std::vector< std::vector<KDL::Vector>::const_iterator > iterators(MAX_SIMPLIFICATION_DEPTH, points.begin());
-
-    int depth = 0;
-
-    do
-    {
-        for (std::vector<PoeTerm>::iterator it = poeTerms.begin(); it != poeTerms.end(); ++it)
-        {
-            it->simplified = false;
-        }
-
-        simplify(depth);
-
-        ScrewTheoryIkSubproblem * subproblem = trySolve(depth);
-
-        if (subproblem != NULL)
-        {
-            steps.push_back(subproblem);
-            iterators.assign(iterators.size(), points.begin());
-            testPoints[0] = points[0];
-            depth = 0;
-            continue;
-        }
-
-        for (int stage = 0; stage < iterators.size(); stage++)
-        {
-            ++iterators[stage];
-
-            if (iterators[stage] == points.end())
-            {
-                iterators[stage] = points.begin();
-                testPoints[stage] = *iterators[stage];
-
-                if (stage == depth)
-                {
-                    depth++;
-                    break;
-                }
-            }
-            else
-            {
-                testPoints[stage] = *iterators[stage];
-                break;
-            }
-        }
-    }
-    while (depth < MAX_SIMPLIFICATION_DEPTH && std::count_if(poeTerms.begin(), poeTerms.end(), unknownTerm) != 0);
-
-    return steps;
-}
-
-// -----------------------------------------------------------------------------
-
-ScrewTheoryIkSubproblem * ScrewTheoryIkProblemBuilder::trySolve(int depth)
-{
-    int unknownsCount = std::count_if(poeTerms.begin(), poeTerms.end(), unknownNotSimplifiedTerm);
-
-    if (unknownsCount == 0 || unknownsCount > MAX_SIMPLIFICATION_DEPTH)
-    {
-        return NULL;
-    }
-
-    std::vector<PoeTerm>::reverse_iterator lastUnknown = std::find_if(poeTerms.rbegin(), poeTerms.rend(), unknownNotSimplifiedTerm);
-    int lastExpId = std::distance(poeTerms.begin(), lastUnknown.base()) - 1;
-    const MatrixExponential & lastExp = poe.exponentialAtJoint(lastExpId);
-
-    if (unknownsCount == 1)
-    {
-        if (depth == 0)
-        {
-            if (lastExp.getMotionType() == MatrixExponential::ROTATION
-                    && !liesOnAxis(lastExp, testPoints[0]))
-            {
-                poeTerms[lastExpId].known = true;
-                return new PadenKahanOne(lastExpId, lastExp, testPoints[0]);
-            }
-
-            if (lastExp.getMotionType() == MatrixExponential::TRANSLATION)
-            {
-                poeTerms[lastExpId].known = true;
-                return new PardosOne(lastExpId, lastExp, testPoints[0]);
-            }
-        }
-
-        if (depth == 1)
-        {
-            if (lastExp.getMotionType() == MatrixExponential::ROTATION
-                    && !liesOnAxis(lastExp, testPoints[0])
-                    && !liesOnAxis(lastExp, testPoints[1]))
-            {
-                poeTerms[lastExpId].known = true;
-                return new PadenKahanThree(lastExpId, lastExp, testPoints[0], testPoints[1]);
-            }
-
-            if (lastExp.getMotionType() == MatrixExponential::TRANSLATION)
-            {
-                poeTerms[lastExpId].known = true;
-                return new PardosThree(lastExpId, lastExp, testPoints[0], testPoints[1]);
-            }
-        }
-    }
-    else if (unknownsCount == 2 && lastUnknown != poeTerms.rend())
-    {
-        std::vector<PoeTerm>::reverse_iterator nextToLastUnknown = lastUnknown;
-        std::advance(nextToLastUnknown, 1);
-
-        if (!unknownNotSimplifiedTerm(*nextToLastUnknown))
-        {
-            return NULL;
-        }
-
-        int nextToLastExpId = lastExpId - 1;
-        const MatrixExponential & nextToLastExp = poe.exponentialAtJoint(nextToLastExpId);
-
-        if (depth == 0)
-        {
-            KDL::Vector r;
-
-            if (lastExp.getMotionType() == MatrixExponential::ROTATION
-                    && nextToLastExp.getMotionType() == MatrixExponential::ROTATION
-                    && !parallelAxes(lastExp, nextToLastExp)
-                    && intersectingAxes(lastExp, nextToLastExp, r))
-            {
-                poeTerms[lastExpId].known = poeTerms[nextToLastExpId].known = true;
-                return new PadenKahanTwo(nextToLastExpId, lastExpId, nextToLastExp, lastExp, testPoints[0], r);
-            }
-
-            if (lastExp.getMotionType() == MatrixExponential::TRANSLATION
-                    && nextToLastExp.getMotionType() == MatrixExponential::TRANSLATION
-                    && !parallelAxes(lastExp, nextToLastExp)
-                    && intersectingAxes(lastExp, nextToLastExp, r))
-            {
-                poeTerms[lastExpId].known = poeTerms[nextToLastExpId].known = true;
-                return new PardosTwo(nextToLastExpId, lastExpId, nextToLastExp, lastExp, testPoints[0]);
-            }
-
-            if (lastExp.getMotionType() == MatrixExponential::ROTATION
-                    && nextToLastExp.getMotionType() == MatrixExponential::ROTATION
-                    && parallelAxes(lastExp, nextToLastExp)
-                    && !colinearAxes(lastExp, nextToLastExp))
-            {
-                poeTerms[lastExpId].known = poeTerms[nextToLastExpId].known = true;
-                return new PardosFour(nextToLastExpId, lastExpId, nextToLastExp, lastExp, testPoints[0]);
-            }
-        }
-    }
-
-    return NULL;
-}
-
-// -----------------------------------------------------------------------------
-
-void ScrewTheoryIkProblemBuilder::simplify(int depth)
-{
-    simplifyWithPadenKahanOne(testPoints[0]);
-
-    if (depth == 1)
-    {
-        simplifyWithPadenKahanThree(testPoints[1]);
-    }
-    else
-    {
-        for (int i = 0; i < poe.size(); i++)
-        {
-            if (poe.exponentialAtJoint(i).getMotionType() == MatrixExponential::TRANSLATION)
-            {
-                simplifyWithPardosOne();
-                break;
-            }
-        }
-    }
-}
-
-// -----------------------------------------------------------------------------
-
-void ScrewTheoryIkProblemBuilder::simplifyWithPadenKahanOne(const KDL::Vector & point)
-{
-    std::vector<PoeTerm>::reverse_iterator ritUnknown = std::find_if(poeTerms.rbegin(), poeTerms.rend(), unknownTerm);
-
-    for (std::vector<PoeTerm>::reverse_iterator rit = ritUnknown; rit != poeTerms.rend(); ++rit)
-    {
-        int i = std::distance(rit, poeTerms.rend()) - 1;
-        const MatrixExponential & exp = poe.exponentialAtJoint(i);
-
-        if (exp.getMotionType() == MatrixExponential::ROTATION && liesOnAxis(exp, point))
-        {
-            rit->simplified = true;
-        }
-        else
-        {
-            break;
-        }
-    }
-}
-
-// -----------------------------------------------------------------------------
-
-void ScrewTheoryIkProblemBuilder::simplifyWithPadenKahanThree(const KDL::Vector & point)
-{
-    std::vector<PoeTerm>::iterator itUnknown = std::find_if(poeTerms.begin(), poeTerms.end(), unknownTerm);
-
-    for (std::vector<PoeTerm>::iterator it = itUnknown; it != poeTerms.end(); ++it)
-    {
-        int i = std::distance(poeTerms.begin(), it);
-        const MatrixExponential & exp = poe.exponentialAtJoint(i);
-
-        if (exp.getMotionType() == MatrixExponential::ROTATION && liesOnAxis(exp, point))
-        {
-            it->simplified = true;
-        }
-        else
-        {
-            break;
-        }
-    }
-}
-
-// -----------------------------------------------------------------------------
-
-void ScrewTheoryIkProblemBuilder::simplifyWithPardosOne()
-{
-    std::vector<PoeTerm>::iterator itUnknown = std::find_if(poeTerms.begin(), poeTerms.end(), unknownNotSimplifiedTerm);
-    std::vector<PoeTerm>::reverse_iterator ritUnknown = std::find_if(poeTerms.rbegin(), poeTerms.rend(), unknownNotSimplifiedTerm);
-
-    int idStart = std::distance(poeTerms.begin(), itUnknown);
-    int idEnd = std::distance(ritUnknown, poeTerms.rend()) - 1;
-
-    if (idStart >= idEnd)
-    {
-        return;
-    }
-
-    const MatrixExponential & firstExp = poe.exponentialAtJoint(idStart);
-    const MatrixExponential & lastExp = poe.exponentialAtJoint(idEnd);
-
-    if (firstExp.getMotionType() == MatrixExponential::TRANSLATION)
-    {
-        for (int i = idEnd - 1; i >= idStart; i--)
-        {
-            const MatrixExponential & nextExp = poe.exponentialAtJoint(i + 1);
-
-            if (i != idStart)
-            {
-                const MatrixExponential & currentExp = poe.exponentialAtJoint(i);
-
-                if (planarMovement(currentExp, nextExp))
-                {
-                    continue;
-                }
-            }
-            else if (normalPlaneMovement(firstExp, nextExp))
-            {
-                for (int j = idStart + 1; j <= idEnd; j++)
-                {
-                    poeTerms[j].simplified = true;
-                }
-            }
-
-            break;
-        }
-    }
-    else if (lastExp.getMotionType() == MatrixExponential::TRANSLATION)
-    {
-        for (int i = idStart + 1; i <= idEnd; i++)
-        {
-            const MatrixExponential & prevExp = poe.exponentialAtJoint(i - 1);
-
-            if (i != idEnd)
-            {
-                const MatrixExponential & currentExp = poe.exponentialAtJoint(i);
-
-                if (planarMovement(prevExp, currentExp))
-                {
-                    continue;
-                }
-            }
-            else if (normalPlaneMovement(prevExp, lastExp))
-            {
-                for (int j = idEnd - 1; j >= idStart; j--)
-                {
-                    poeTerms[j].simplified = true;
-                }
-            }
-
-            break;
-        }
-    }
 }
 
 // -----------------------------------------------------------------------------
